Add From<Fault> for Failure
This lets code avoid writing `Failure::Fault(...)` a lot, by exploiting
the fact that the ? operator invokes a From impl if available.

There currently aren't a lot of opportunities to use this in HIF itself
because the HIF code heavily prefers "return Err(...)" to using ?.
However, it's making my Hiffy changes easier.